add unique file structure and fix project errors

+++ src/controllers/api/v1/brand_controller.py
@@ -1,15 +1,18 @@
+import os
 from typing import Annotated
 
-from fastapi import APIRouter, status, Depends, HTTPException, Query
 from sqlalchemy.orm import Session
 from fastapi_pagination import Page, Params
 from fastapi_pagination.ext.sqlalchemy import paginate
+from fastapi import APIRouter, status, Depends, HTTPException, Query
 
-from src.schemas.brand_schema import CreateBrand, ResponseBrand, create_brand_form
 from src.config.database import get_db
 from src.models.brand_model import BrandModel
-from src.validators.brand_validator import validate_unique_email_and_phone
+from src.models.suplier_model import SuplierModel
 from src.filters.brand_filters import BrandFilters
+from src.models.product_template_model import ProductTemplateModel
+from src.validators.brand_validator import validate_unique_email_and_phone
+from src.schemas.brand_schema import CreateBrand, ResponseBrand, create_brand_form
 
 
 router = APIRouter()
@@ -34,6 +37,7 @@ def create_brand(data : CreateBrand = Depends(validate_unique_email_and_phone), 
         return brand
     
     except Exception:
+        os.remove(data.logo)
         db.rollback()
         raise
 
@@ -54,7 +58,7 @@ def read_brand(filters : Annotated[BrandFilters, Query()] = None, db : Session =
 
 
 @router.put("/{id}", response_model=ResponseBrand, status_code=status.HTTP_200_OK)
-def update_brand(id : int, data : CreateBrand = Depends(validate_unique_email_and_phone),  db : Session = Depends(get_db)):
+def update_brand(id : int, data : CreateBrand = Depends(create_brand_form),  db : Session = Depends(get_db)):
     brand = db.query(BrandModel).filter(BrandModel.id == id).first()
 
     if not brand:
@@ -62,6 +66,9 @@ def update_brand(id : int, data : CreateBrand = Depends(validate_unique_email_an
             status_code = status.HTTP_404_NOT_FOUND,
             detail = "Data not found"
         )
+
+    file_path = brand.logo
+    os.remove(f"{file_path}")
 
     update_data = data.model_dump()
     for key, value in update_data.items():
@@ -74,6 +81,7 @@ def update_brand(id : int, data : CreateBrand = Depends(validate_unique_email_an
         return brand
     
     except Exception:
+        os.remove(data.logo)
         db.rollback()
         raise
 
@@ -88,11 +96,32 @@ def delete_brand(id : int, db : Session = Depends(get_db)):
             status_code = status.HTTP_404_NOT_FOUND,
             detail = "Data not found"
         )
+    
+
+    product_template = db.query(ProductTemplateModel).filter(ProductTemplateModel.brand_id == id).first()
+
+    if product_template:
+        raise HTTPException(
+            status_code = status.HTTP_409_CONFLICT,
+            detail = "In Product Template, Brand is exists"
+        )
+    
+
+    suplier = db.query(SuplierModel).filter(SuplierModel.brand_id == id).first()
+    
+    if suplier:
+        raise HTTPException(
+            status_code = status.HTTP_409_CONFLICT,
+            detail = "In Suplier, Brand is exists"
+        )
+    
+
+    file_path = brand.logo
+    os.remove(f"{file_path}")
 
     try:
         db.delete(brand)
         db.commit()
-        db.refresh(brand)
 
         return {"status" : "Data successfuly delete"}
