+++ src/controllers/api/v1/sub_category_controller.py
@@ -1,16 +1,18 @@
-from typing import List, Annotated
+import os
+from typing import Annotated
 
-from fastapi import APIRouter, status, Depends, HTTPException, Query
 from sqlalchemy.orm import Session
 from fastapi_pagination import Page, Params
 from fastapi_pagination.ext.sqlalchemy import paginate
+from fastapi import APIRouter, status, Depends, HTTPException, Query
 
-from src.schemas.sub_category_schema import CreateSubCategory, ResponseSubCategory, create_sub_category_form
-from src.validators.sub_category_validator import validate_unique_code
 from src.config.database import get_db
-from src.models.sub_category_model import SubCategoryModel
 from src.models.category_model import CategoryModel
+from src.models.sub_category_model import SubCategoryModel
 from src.filters.sub_category_filters import SubCategoryFilter
+from src.models.product_template_model import ProductTemplateModel
+from src.validators.sub_category_validator import validate_unique_code
+from src.schemas.sub_category_schema import CreateSubCategory, ResponseSubCategory, create_sub_category_form
 
 
 router = APIRouter()
@@ -43,6 +45,7 @@ def create_sub_category(data : CreateSubCategory = Depends(validate_unique_code)
         return sub_category
 
     except Exception:
+        os.remove(data.icon)
         db.rollback()
         raise
 
@@ -63,7 +66,7 @@ def read_sub_category(filters : Annotated[SubCategoryFilter, Query()] = None, db
 
 
 @router.put("/{id}", response_model=ResponseSubCategory, status_code=status.HTTP_200_OK)
-def update_sub_category(id: int, data : CreateSubCategory = Depends(validate_unique_code), db : Session = Depends(get_db)):
+def update_sub_category(id: int, data : CreateSubCategory = Depends(create_sub_category_form), db : Session = Depends(get_db)):
     sub_category = db.query(SubCategoryModel).filter(SubCategoryModel.id == id).first()
 
     if not sub_category:
@@ -80,8 +83,10 @@ def update_sub_category(id: int, data : CreateSubCategory = Depends(validate_uni
             detail = "Category not found"
         )
 
+    file_path = sub_category.icon
+    os.remove(f"{file_path}")
+
     update_data = data.model_dump(exclude={"category_name"})
-    
     for key,value in update_data.items():
         setattr(sub_category,key,value)
 
@@ -94,6 +99,7 @@ def update_sub_category(id: int, data : CreateSubCategory = Depends(validate_uni
         return sub_category
 
     except Exception:
+        os.remove(data.icon)
         db.rollback()
         raise
 
@@ -108,6 +114,17 @@ def delete_sub_category(id : int, db : Session = Depends(get_db)):
             status_code = status.HTTP_404_NOT_FOUND,
             detail = "Data Not Found"
         )
+
+    product_template = db.query(ProductTemplateModel).filter(ProductTemplateModel.sub_category == id).first()
+    
+    if product_template:
+        raise HTTPException(
+            status_code = status.HTTP_409_CONFLICT,
+            detail = "In Product Template, Sub Category is exists"
+        )
+
+    file_path = sub_category.icon
+    os.remove(f"{file_path}")
 
     try:
         db.delete(sub_category)
